レシピ画面のサブヘッダー修正
