+++ components/organisms/SubHeader/SubHeader.vue
@@ -3,17 +3,17 @@
     <bread-crumb
       class="content_header_bread-crumb"
       :link="`/home/food/`"
-      list-name="食材"
+      :list-name="listName"
     >
       <span slot="link-name">マイページ</span>
     </bread-crumb>
 
     <tab-link
+      :is-values="isValues"
       :is-register-active="isRegisterActive"
       :is-view-active="isViewActive"
-      :param-id-page="`/home/food/${food.paramId}`"
-      :register-page="`/home/food/register`"
-      :to="`/home/food/${food.paramId}`"
+      :param-id-page="paramIdPage"
+      :register-page="registerPage"
     >
       <template slot="btn">
         <slot name="btn" />
@@ -32,11 +32,27 @@ export default {
     TabLink
   },
   props: {
+    listName: {
+      type: String,
+      default: null
+    },
     isViewActive: {
       type: Boolean
     },
     isRegisterActive: {
       type: Boolean
+    },
+    isValues: {
+      type: Array,
+      default: null
+    },
+    paramIdPage: {
+      type: String,
+      required: true
+    },
+    registerPage: {
+      type: String,
+      required: true
     }
   },
   data() {
@@ -47,9 +63,9 @@ export default {
     }
   },
   created() {
-    if (this.$store.getters['food/foods'].length !== 0) {
-      this.food.paramId = this.$store.getters['food/foods'][0].id
-    }
+    // if (this.$store.getters['food/foods'].length !== 0) {
+    //   this.food.paramId = this.$store.getters['food/foods'][0].id
+    // }
   }
 }
 </script>
